henry/my-setera-BACKEND/requirements.txt
Django==4.1.
django-celery-beat==2.4.0
django-celery-results==2.4.0
djangorestframework==3.14.0
django-rest-swagger==2.2.0
drf-yasg==1.21.4
django-rest-swagger==2.2.
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
pyotp==2.8.0
Pillow==9.4.0
redis==4.4.2
celery==5.2.7
django-rest-passwordreset==1.3.0
django-cors-headers==3.13.0
django-seed==0.3.1
django-rq==2.7.0
django-redis==5.2.0
django-environ==0.10.0
whitenoise==6.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.5
django_extensions==3.2.1
django-filter==23.1

